bug fixes, update models
